+++ demo/DemoV1b/Model/L2User.generated.cs
@@ -77,6 +77,7 @@ namespace DemoV1b.Model {
         [JsonProperty("ageGroup")]
         public virtual int? AgeGroup { get; set; } = 40;
 
+        [Obsolete("do not use this")]
         [JsonConverter(typeof(CustomTimeConverter))]
         [JsonProperty("preferredContactTime")]
         public virtual TimeOnly PreferredContactTime { get; set; } = TimeOnly.Parse("12:00");